# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  api.ApiResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  api.addMemberRequest:
    properties:
      member_id:
        type: integer
    required:
    - member_id
    type: object
...
  api.createGroupRequest:
    properties:
      name:
        type: string
      notice:
        type: string
    required:
    - name
    type: object
  api.createGroupResponse:
    properties:
      group_id:
        type: integer
    type: object
...
  api.updateGroupRequest:
    properties:
      name:
        type: string
      notice:
        type: string
    required:
    - name
    type: object
  api.updateMemberRoleRequest:
    properties:
      role:
        type: integer
    required:
    - role
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: GoIM 即时通讯系统 API — 支持私聊、群聊、朋友圈动态。
  title: GoIM API
  version: "1.0"
paths:
...
  /group:
    post:
      consumes:
      - application/json
      description: 创建一个新群组，创建者自动成为群主
      parameters:
      - description: 群组信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.createGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/api.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/api.createGroupResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/api.ApiResponse'
      security:
      - BearerAuth: []
      summary: 创建群组
      tags:
      - 群组
  /group/{groupID}:
    get:
      consumes:
      - application/json
      description: 根据群组ID获取群组详细信息
      parameters:
      - description: 群组ID
        format: int64
        in: path
        name: groupID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/api.ApiResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "404":
          description: 群组不存在
          schema:
            $ref: '#/definitions/api.ApiResponse'
      security:
      - BearerAuth: []
      summary: 获取群组信息
      tags:
      - 群组
    put:
      consumes:
      - application/json
      description: 更新群组的名称和公告，仅群主或管理员可操作
      parameters:
      - description: 群组ID
        format: int64
        in: path
        name: groupID
        required: true
        type: integer
      - description: 更新群组信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.updateGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "404":
          description: 群组不存在
          schema:
            $ref: '#/definitions/api.ApiResponse'
      security:
      - BearerAuth: []
      summary: 更新群组信息
      tags:
      - 群组
  /group/{groupID}/leave:
    post:
      consumes:
      - application/json
      description: 当前用户退出指定群组，群主不能直接退群（需先转让群主）
      parameters:
      - description: 群组ID
        format: int64
        in: path
        name: groupID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 退出成功
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "403":
          description: 群主不能退群
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "404":
          description: 群组不存在
          schema:
            $ref: '#/definitions/api.ApiResponse'
      security:
      - BearerAuth: []
      summary: 退出群组
      tags:
      - 群组
  /group/{groupID}/member:
    post:
      consumes:
      - application/json
      description: 向群组添加一名新成员，仅群主或管理员可操作
      parameters:
      - description: 群组ID
        format: int64
        in: path
        name: groupID
        required: true
        type: integer
      - description: 成员信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.addMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 添加成功
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "404":
          description: 群组不存在
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "409":
          description: 已是成员或群组已满
          schema:
            $ref: '#/definitions/api.ApiResponse'
      security:
      - BearerAuth: []
      summary: 添加群成员
      tags:
      - 群组
  /group/{groupID}/member/{memberID}:
    delete:
      consumes:
      - application/json
      description: 从群组中移除一名成员；若移除自己则为退群，移除他人则为踢出（仅群主/管理员可踢人）
      parameters:
      - description: 群组ID
        format: int64
        in: path
        name: groupID
        required: true
        type: integer
      - description: 成员ID
        format: int64
        in: path
        name: memberID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 移除成功
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "404":
          description: 群组不存在
          schema:
            $ref: '#/definitions/api.ApiResponse'
      security:
      - BearerAuth: []
      summary: 移除群成员
      tags:
      - 群组
  /group/{groupID}/member/{memberID}/role:
    put:
      consumes:
      - application/json
      description: 更新群组成员角色（普通成员/管理员），仅群主可操作
      parameters:
      - description: 群组ID
        format: int64
        in: path
        name: groupID
        required: true
        type: integer
      - description: 成员ID
        format: int64
        in: path
        name: memberID
        required: true
        type: integer
      - description: 角色信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.updateMemberRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "404":
          description: 群组不存在
          schema:
            $ref: '#/definitions/api.ApiResponse'
      security:
      - BearerAuth: []
      summary: 更新成员角色
      tags:
      - 群组
  /group/{groupID}/members:
    get:
      consumes:
      - application/json
      description: 获取指定群组的所有成员信息
      parameters:
      - description: 群组ID
        format: int64
        in: path
        name: groupID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/api.ApiResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/api.ApiResponse'
      security:
      - BearerAuth: []
      summary: 获取群成员列表
      tags:
      - 群组
...
securityDefinitions:
  BearerAuth:
    description: 在 Authorization 头中携带 JWT 访问令牌，格式：Bearer <token>
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"